Move CircuitNode and builder into own file
Moves CircuitNode into store::circuit_node::{CircuitNode, CircuitNodeBuilder}

Structs are still exposed only in store:: via pub use. This change
also requires getters to be implemented on the structs to be used
throughout the store implementations.
