+++ libsplinter/src/admin/store/builders.rs
@@ -18,7 +18,7 @@ use crate::admin::messages::{is_valid_circuit_id, is_valid_service_id};
 
 use super::error::BuilderError;
 use super::{
-    AuthorizationType, CircuitNode, CircuitProposal, DurabilityType, PersistenceType, ProposalType,
+    AuthorizationType, CircuitProposal, DurabilityType, PersistenceType, ProposalType,
     ProposedCircuit, ProposedNode, ProposedService, RouteType, VoteRecord,
 };
 
@@ -437,70 +437,6 @@ impl CircuitProposalBuilder {
     }
 }
 
-/// Builder for creating a `CircutNode`
-#[derive(Default, Clone)]
-pub struct CircuitNodeBuilder {
-    node_id: Option<String>,
-    endpoints: Option<Vec<String>>,
-}
-
-impl CircuitNodeBuilder {
-    /// Creates a `CircuitNodeBuilder`
-    pub fn new() -> Self {
-        CircuitNodeBuilder::default()
-    }
-
-    /// Returns the unique node ID
-    pub fn node_id(&self) -> Option<String> {
-        self.node_id.clone()
-    }
-
-    /// Returns the list of endpoints for the node
-    pub fn endpoints(&self) -> Option<Vec<String>> {
-        self.endpoints.clone()
-    }
-
-    /// Sets the node ID
-    ///
-    /// # Arguments
-    ///
-    ///  * `node_id` - The unique node ID for node
-    pub fn with_node_id(mut self, node_id: &str) -> CircuitNodeBuilder {
-        self.node_id = Some(node_id.into());
-        self
-    }
-
-    /// Sets the endpoints
-    ///
-    /// # Arguments
-    ///
-    ///  * `endpoints` - The list of endpoints for the node
-    pub fn with_endpoints(mut self, endpoints: &[String]) -> CircuitNodeBuilder {
-        self.endpoints = Some(endpoints.into());
-        self
-    }
-
-    /// Builds the `CircuitNode`
-    ///
-    /// Returns an error if the node ID or endpoints are not set
-    pub fn build(self) -> Result<CircuitNode, BuilderError> {
-        let node_id = self
-            .node_id
-            .ok_or_else(|| BuilderError::MissingField("node_id".to_string()))?;
-
-        let endpoints = self
-            .endpoints
-            .ok_or_else(|| BuilderError::MissingField("endpoints".to_string()))?;
-
-        let node = CircuitNode {
-            id: node_id,
-            endpoints,
-        };
-
-        Ok(node)
-    }
-}
-
 /// Builder for creating a `ProposedNode`
 #[derive(Default, Clone)]
 pub struct ProposedNodeBuilder {
@@ -559,15 +495,6 @@ impl ProposedNodeBuilder {
         let node = ProposedNode { node_id, endpoints };
 
         Ok(node)
-    }
-}
-
-impl From<ProposedNode> for CircuitNode {
-    fn from(node: ProposedNode) -> Self {
-        CircuitNode {
-            id: node.node_id,
-            endpoints: node.endpoints,
-        }
     }
 }
